src/3mw_backend_service/backend/api/routes.py
from rest_framework import routers
from .views import PlantViewSet, DatapointViewSet

router = routers.DefaultRouter()
router.register(r'plants', PlantViewSet)
router.register(r'datapoints', DatapointViewSet)